requirements.txt
RPi.GPIO==0.7.0
pyaudio==0.2.13
requests==2.31.0
